Changes:
• Added set methods to the IniFileStream systems.
• Adding in Directory class.
• Fixed casting warning in allocators.
• Fixed possible invalid deallocation in Managed Pointers because of the Cast method.
• Managed Pointers now rely on reflection for deallocation.
• Managed Pointers only accept objects with a GetType function that returns a Type. 
• Resources now load meta data with LoadMeta and SaveMeta methods. They only read/write to IniFileStreams however. Saving will be done through the ResourceDatabase class.
• LoadTexture has now been implemented in resource database and is ready for testing.
• UnloadResource worked for LoadTexture on a single resource.

+++ GameEngine_Project/Common/Memory/FrameAllocator.cpp
@@ -92,7 +92,7 @@ namespace Engine
             ///Write header info
             MemoryHeader * header = (MemoryHeader*)(reinterpret_cast<void*>(alignedAddress - sizeof(MemoryHeader)));
             UInt8 flags = MemoryFlags::VALID | (m_ID = 0 ? MemoryFlags::STATIC : MemoryFlags::DYNAMIC) | MemoryFlags::FRAME;
-            UInt16 objectSize = aSize;
+            UInt16 objectSize = (UInt16)aSize;
             header->Write(flags, m_ID,objectSize);
             //header->bytes[MemoryHeader::ALLOCATOR_ID] = m_ID;
             //header->bytes[MemoryHeader::OFFSET] = adjustment;
